requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
reportlab>=4.0.0
